Mention .lslo or .lsl; add preparation instructions for AVpos.
There is a .lsl file (the Lockguard plugin) and there may be more to come, so mention ".lslo or .lsl" instead of just ".lslo".

One user in the group had problems because the AVpos notecard had never been saved; to prevent that, add instructions on preparing an empty AVpos notecard that has been saved.