Check permissions before starting/stopping overlay animations
This partially solves a race condition when avatars revoke animation permissions on stand, but the message needs to travel through several scripts and, in case of extreme lag conditions, can attempt to stop animations after the revocation arrives, resulting in an error.

It still suffers a TOCTTOU race, as demonstrated by the reporter. That race is probably unsolvable.

Fixes #119. Big thanks to @ohhmye for reporting and investigating this bug.