Hardcode msg string to 'Restricted'
Sei-Lisa suggested a change that would no longer send a hardcoded message string to AVsitter. I understood the idea behind this and happily implemented the change, and then I haven't thoroughly tested it. Now I realized that it broke the functionality of the script entirely. The thing is that MENU and BUTTON may not have the same name otherwise AVsitter will compute the request internally instead of passing it via link message. The restricted menu would always be available to everyone. I have now reverted this change, which will fix this plugin.